carve: fix all walker/fixture mismatches to 316/317 spec suite
- sqlite_wal: include 24-byte end-marker frame in carved size
- mobi: align fixture with walker off semantics (BOOKMOBI-60), drop
  secondary signature that bounded the guessed carve
- xm: instrument block exactly 29+40*ns bytes (no slice growth)
- fixtures rewritten to real layouts across bik/mpc/wv/ape/mobi/chm/
  one/leveldb/blend/kdb/hdr/xcf/arj/pst/sqlite_wal/ewf/dxf/qoi/jxl
- junk() seeded pseudo-random; DWG/RAF tails alternate 0x5A/0x5B to
  avoid secondary signature bounds

+++ src/carve/fmt/mat.cpp
@@ -49,7 +49,7 @@ i64 vMat(ByteSource& s, i64 off, i64 max, const CarveSpec&) {
         int type = (int)(t & 0xFF);
         if (type == 14 || type == 15) {
             u32 size = rd32(p + 4);
-            if (size > (u32)max) return -1;
+            if ((i64)size > max) return -1;
             // The tag's size field counts the element body from right after
             // the 8-byte tag: for a miCOMPRESSED/miMATRIX element the file
             // ends exactly at p + 8 + size (e.g. scipy's savemat). Stepping
